+++ src/ska_low_mccs_daq/daq_receiver/daq_receiver_device.py
@@ -413,6 +413,7 @@ class MccsDaqReceiver(MccsBaseDevice):
             ("Stop", "stop_daq"),
             ("StartDataRateMonitor", "start_data_rate_monitor"),
             ("StopDataRateMonitor", "stop_data_rate_monitor"),
+            ("MarkDone", "mark_scan_done"),
         ]:
             self.register_command_object(
                 command_name,
@@ -1101,14 +1102,18 @@ class MccsDaqReceiver(MccsBaseDevice):
         (result_code, message) = handler(interval)
         return ([result_code], [message])
 
-    @command
-    def CallComponentCallback(self, argin: str) -> None:
+    @command(dtype_out="DevVarLongStringArray")
+    def MarkDone(self: MccsDaqReceiver) -> DevVarLongStringArrayType:
         """
-        Patched method to call component callback directly.
+        Change the data dictionary to mark the end of a scan.
 
-        :param argin: json-ified dict to call component callback with.
+        :return: A tuple containing a return code and a string
+            message indicating status. The message is for
+            information purpose only.
         """
-        self._component_state_callback(**json.loads(argin))
+        handler = self.get_command_object("MarkDone")
+        (result_code, message) = handler()
+        return ([result_code], [message])
 
     # ----------
     # Attributes
@@ -1411,6 +1416,34 @@ class MccsDaqReceiver(MccsBaseDevice):
         :return: a string corresponding to the library filename.
         """
         return self.component_manager.daq_library
+
+    @attribute(
+        dtype="DevBoolean",
+        doc=("If the last directory change was successful"),
+    )
+    def directoryChangeResult(self: MccsDaqReceiver) -> bool:
+        """
+        Get the last result of a directory tag command.
+
+        :return: a bool corresponding to the last result of a directory tag change
+        """
+        return self.component_manager.directory_change_result
+
+    @attribute(
+        dtype="DevString",
+        doc=("Current status of a scan",),
+    )
+    def scanStatus(self: MccsDaqReceiver) -> str:
+        """
+        Get the current status of recieving a scan.
+
+        :return: a string stating if we are receiving a scan or not.
+            "Active" if the we are curently recording a scan.
+            "Inactive" if the receiver is stopped.
+        """
+        if self.component_manager._scan_in_progress:
+            return "Active"
+        return "Inactive"
 
 
 # ----------